Refactor CityController

+++ src/main/java/dev/webservices/customerapi/Controller/CityController.java
@@ -13,7 +13,9 @@ import org.springframework.web.bind.annotation.RequestParam;
 import org.springframework.web.bind.annotation.RestController;
 
 import dev.webservices.customerapi.Entity.City;
+import dev.webservices.customerapi.Entity.Country;
 import dev.webservices.customerapi.Service.CityService;
+import dev.webservices.customerapi.Service.CountryService;
 
 @RestController
 @RequestMapping("/city")
@@ -21,6 +23,9 @@ public class CityController {
 
     @Autowired
     private CityService cityService;
+
+    @Autowired
+    private CountryService countryService;
 
     // Save city in the database
     @PostMapping("/")
@@ -32,15 +37,22 @@ public class CityController {
     // Find city by ID
     @GetMapping("/")
     public City getCity(@RequestParam Long id) {
-        return cityService.findById(id).get();
+
+        // Check if the city with the given ID present
+        Optional<City> cityId = cityService.findById(id);
+        if (cityId.isPresent()) {
+            return cityService.findById(id).get();
+        } else {
+            return null;
+        }
     }
 
     // Update city data
     @PutMapping("/")
     public String update(@RequestBody City city) {
-        Optional<City> cityId = cityService.findById(city.getId());
 
         // Check if the city with the given ID present
+        Optional<City> cityId = cityService.findById(city.getId());
         if (cityId.isPresent()) {
             cityService.update(city);
             return "City updated successfully!";
@@ -53,8 +65,52 @@ public class CityController {
     // Delete city
     @DeleteMapping("/")
     public String delete(@RequestParam Long id) {
-        cityService.delete(id);
-        return "City record deleted successfullt!";
+
+        // Check if the city with the given ID present
+        Optional<City> cityId = cityService.findById(id);
+        if (cityId.isPresent()) {
+            cityService.delete(id);
+            return "City record deleted successfully!";
+        } else {
+            return "City not found!";
+        }
     }
+
+    // // Find country by ID
+    // @GetMapping("/")
+    // public void findCountry(@RequestParam Long id) {
+    // countryService.findById(id);
+    // System.out.println(cityService.toString());
+    // }
+
+    // // Update country
+    // @PutMapping("/")
+    // public void updateCountry(@RequestParam Long id) {
+
+    // Optional<Country> country = countryService.findById(id);
+
+    // // Check if Country with given ID exist
+    // if (country.isPresent()) {
+    // countryService.update(id);
+    // System.out.println("Country updated successfully!");
+    // } else {
+    // System.out.println("Country not found!");
+    // }
+    // }
+
+    // // Delete country
+    // @DeleteMapping("/")
+    // public void deleteCountry(@RequestParam Long id) {
+
+    // Optional<Country> country = countryService.findById(id);
+
+    // // Check if Country with given ID exist
+    // if (country.isPresent()) {
+    // countryService.delete(id);
+    // System.out.println("Country record deleted successfully!");
+    // } else {
+    // System.out.println("Country not found!");
+    // }
+    // }
 
 }
